Single Target Texture
I'm was able to get the target texture working but I don't know how to
get the texture pointer I create in GameState.cpp to be available in
Target.cpp. If the target class was inside the gamestate class this
would be a nonissue - do you know how we could make this work while
keeping the Target.cpp class seperate?